requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.8.3
click==8.1.8
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.35.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.44.0
typer==0.16.1
typing_extensions==4.13.2
urllib3==2.2.3
uvicorn==0.33.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==13.1
